pyproject.toml
[tool.poetry]
name = "kaggle-home-credit"
version = "0.1.0"
description = ""
authors = ["Wojtek Ostrowski <wojciech.ostrowski@esky.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13.0"
jupyter = "^1.0.0"
numpy = "^1.26.4"
pandas = "^2.2.1"
scikit-learn = "^1.4.1.post1"
scipy = "^1.12.0"
tqdm = "^4.66.2"
lightgbm = "^4.3.0"
catboost = "^1.2.3"
xgboost = "^2.0.3"
matplotlib = "^3.8.3"
seaborn = "^0.13.2"
polars = "^0.20.16"
mlflow = "^2.11.3"
hyperopt = "^0.2.7"
feature-engine = "^1.7.0"
notebook = "^7.1.3"
optuna = "^3.6.1"
pytest = "^8.2.2"
python-dotenv = "^1.0.1"
kfp = "^2.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
